backend/app/core/config.py
import os
from pydantic_settings import BaseSettings


class Settings(BaseSettings):
    OPENAI_API_KEY: str
    SUPABASE_URL: str
    SUPABASE_KEY: str
    ENVIRONMENT: str

    class Config:
        env_file = ".env"
        env_file_encoding = "utf-8"


settings = Settings()
os.environ["OPENAI_API_KEY"] = settings.OPENAI_API_KEY
