backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Julia, Ashley, Fatima, Helen"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
phidata = "^2.5.30"
openai = "^1.54.3"
supabase = "^2.10.0"
uvicorn = "^0.32.0"
fastapi = "^0.115.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.poetry.scripts]
start = "app.main:start"
